20260725: grade sheet — enforce outcome_criteria score bounds server-side; one chrome band; unique cell testids
SERVER-SIDE BOUNDS (the important one). outcome_criteria.max_score is already the
authority — all 26 IB-MYP criteria declare max_score=8, and the cell input's
max="8" is correctly derived from it. But the write path only type-checked:
applyValueStrict parsed the raw value as a float and stored it. Saving 76 into an
8-point criterion returned 200 and persisted.

The rule already existed and was unreachable: outcome_evaluation/numeric_score.go
implements exactly this Min/MaxScore check, but nothing outside its own package
references it — not composition, not consumer, not registry. Dead code, the same
shape as the DERIVE use case.

createCell/applyValueStrict now carry cellBounds and gate numerically: min/max
INCLUSIVE on both ends, with an explicit NaN/±Inf reject (they parse as float64
and compare false against every bound). 6 tests: above-max and below-min are
rejected with UpdateTaskOutcome never called; 0/8/5 all save; NaN/Inf rejected.
Verified live as a teacher — 76 now yields data-cg-state=error, aria-invalid, and
"Save failed" with nothing written.

Why it is not cosmetic: aggregation_method is AGGREGATION_METHOD_MAXIMUM, so one
out-of-range value would dominate its criterion's roll-up onto the report card.

ONE CHROME BAND: the GRADING PERIOD APPROVALS block and the Columns/Download
buttons now share a single .om-toolbar row instead of two stacked bands. The
approval block loses its card chrome but keeps both semesters' chips and Submit
buttons; verified at two viewports.

UNIQUE CELL TESTIDS: the old shape used the first 8 chars of the client id, and
every education1 client id is a UUIDv7 sharing the prefix "019ecb8e" — so one
testid resolved to 74 elements, identifying the COLUMN rather than the cell. Now
row-unique. No spec updates needed: grep across apps/school-admin/tests and
fayna tests found zero references to om-cell-*, om-scope*, or om-save.

SCOPE TOGGLE KEPT, deliberately. "My students"/"All students" only renders when
the principal holds workspace:list, so a teacher never sees it — it is an
operator affordance, not a teacher workaround. Removing it would take a
capability away from operators while the sheet is still grade-scoped. It goes
when the class-scoping lands.

Pre-existing failure, not mine: TestPlacement (god-file lint) flags three files
all already over threshold at HEAD.

+++ domain/operation/outcome_matrix/list/page.go
@@ -1275,12 +1275,19 @@ func criteriaIDFromColumnKey(k string) string {
 	return k
 }
 
+// cellTestID composes the per-cell data-testid. It carries the FULL row id (not
+// short()): UUIDv7 row ids share a tenant/time prefix, so short() truncated every
+// row in a workspace to the SAME 8 characters and collapsed a whole COLUMN of
+// cells onto one testid (measured: one selector matched 74 elements) — an E2E
+// could never address a single row's cell. The same F3 fix already landed on the
+// om-row-* and om-note-* families (buildRows above); this is the last member.
+// clientID (unique per row) + colKey (unique per leaf column) ⇒ unique per cell.
 func cellTestID(clientID, colKey string, readOnly bool) string {
 	base := "om-cell-"
 	if readOnly {
 		base = "om-cell-ro-"
 	}
-	return base + short(clientID) + "-" + slug(colKey)
+	return base + clientID + "-" + slug(colKey)
 }
 
 // short truncates an opaque id for a testid suffix.
